gst/goom/: Compile fixes for x86-64.
Original commit message from CVS:
* gst/goom/goomsl_heap.c: (align_it):
* gst/goom/plugin_info.c: (setOptimizedMethods):
Compile fixes for x86-64.

+++ gst/goom/goomsl_heap.c
@@ -46,8 +46,8 @@ align_it (GoomHeap * _this, int alignment)
 {
   if ((alignment > 1) && (_this->number_of_arrays > 0)) {
     void *last_array = _this->arrays[_this->number_of_arrays - 1];
-    int last_address = (int) last_array + _this->consumed_in_last_array;
-    int decal = (last_address % alignment);
+    long last_address = (long) last_array + _this->consumed_in_last_array;
+    long decal = (last_address % alignment);
 
     if (decal != 0) {
       _this->consumed_in_last_array += alignment - decal;
